NOJIRA move SparkSessionProvider into provider package
