+++ src/test/java/uk/gov/justice/digital/domain/DomainExecutorTest.java
@@ -15,7 +15,7 @@ import uk.gov.justice.digital.domain.model.TableDefinition;
 import uk.gov.justice.digital.domain.model.TableInfo;
 import uk.gov.justice.digital.domain.model.TableTuple;
 import uk.gov.justice.digital.exception.DomainExecutorException;
-import uk.gov.justice.digital.job.SparkSessionProvider;
+import uk.gov.justice.digital.provider.SparkSessionProvider;
 import uk.gov.justice.digital.service.DataStorageService;
 import uk.gov.justice.digital.service.SparkTestHelpers;
